refactor(cli/audio): split analyzer submodule into file
